Completed monorepo structure - python folder renamed to py-client - client folder renamed to js-client
