+++ .gitignore
@@ -5,14 +5,14 @@ node_modules
 .idea/
 .next
 .env
-python/__pycache__/
+py-client/__pycache__/
 *.pyc
 *.pyo
 *.egg-info/
 dist/
 build/
-python/dist/
-python/build/
+py-client/dist/
+py-client/build/
 .pytest_cache/
-python/.pytest_cache/
+py-client/.pytest_cache/
 .vscode
