+++ js-client/.npmignore
