+++ py-client/setup.cfg
