+++ server/package.json
@@ -21,6 +21,7 @@
     "@mantine/next": "^6.0.8",
     "@mantine/prism": "^6.0.8",
     "@next/font": "13.1.6",
+    "@rebuff/client": "*",
     "@pinecone-database/pinecone": "^0.0.10",
     "@supabase/auth-helpers-nextjs": "^0.6.0",
     "@supabase/auth-helpers-react": "^0.3.1",
