+++ server/tsconfig.json
@@ -11,14 +11,24 @@
     "jsx": "preserve",
     "incremental": true,
     "outDir": "dist",
-    "rootDir": ".",
+    "rootDir": "./",
     "composite": true,
     "baseUrl": ".",
     "paths": {
+      "@rebuff/types/*": ["../types/src/*"],
+      "@rebuff/client/*": ["../js-client/src/*"],
       "@/*": ["./*"]
     },
     "resolveJsonModule": true
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
-  "exclude": ["node_modules", "dist"]
+  "exclude": ["node_modules", "dist"],
+  "references": [
+    {
+      "path": "../js-client"
+    },
+    {
+      "path": "../types"
+    }
+  ]
 }
